+++ docs/api-reference/arcgis/deck-layer.md
@@ -2,34 +2,37 @@
 
 This class inherits from the ArcGIS [Layer](https://developers.arcgis.com/javascript/latest/api-reference/esri-layers-Layer.html) class and can be added to maps created with the ArcGIS API for JavaScript.
 
-`DeckLayer` is only available when `loadArcGISModules()` is resolved. At the moment, it only supports 2D integration.
+At the moment, `DeckLayer` only supports 2D integration.
 
 ## Usage
 
 ```js
-import {loadArcGISModules} from '@deck.gl/arcgis';
+import {DeckLayer} from '@deck.gl/arcgis';
+import {ScatterplotLayer} from '@deck.gl/layers';
+import ArcGISMap from '@arcgis/core/Map';
+import MapView from '@arcgis/core/views/MapView';
 
-loadArcGISModules([
-  'esri/Map',
-  'esri/views/MapView'
-]).then(({DeckLayer, modules}) => {
-  const [ArcGISMap, MapView] = modules;
+const layer = new DeckLayer({
+  'deck.layers': [
+    new ScatterplotLayer({
+      data: [
+        {position: [0.119, 52.205]}
+      ],
+      getPosition: d => d.position,
+      getColor: [255, 0, 0],
+      radiusMinPixels: 20
+    })
+  ]
+});
 
-  const layer = new DeckLayer({
-    'deck.layers': [
-      // deck.gl layers
-    ]
-  });
-
-  const mapView = new MapView({
-    container: "viewDiv",
-    map: new ArcGISMap({
-      basemap: "dark-gray-vector",
-      layers: [layer]
-    }),
-    center: [0.119, 52.205],
-    zoom: 5
-  });
+const mapView = new MapView({
+  container: "viewDiv",
+  map: new ArcGISMap({
+    basemap: "dark-gray-vector",
+    layers: [layer]
+  }),
+  center: [0.119, 52.205],
+  zoom: 5
 });
 ```
